requirements.txt
opencv-python
numpy
Pillow
tensorflow
matplotlib
pyserial
requests
pygame